Remove btn invitation if user == current_user

+++ app/helpers/application_helper.rb
@@ -17,11 +17,13 @@ module ApplicationHelper
   end
 
   def send_or_recall_friend_request(user)
-    request = Friendship.find_by(user: current_user, friend: user)
-    if request
-      link_to('Recall friend request', user_friendship_path(id: request.id, user_id: current_user.id, friend_id: user.id), method: :delete, class: 'btn-alert')
-    else
-      link_to('Send friend request', user_friendships_path(user_id: current_user.id, friend_id: user.id, confirmed: false), method: :post, class: 'btn-primary')
+    if current_user != user
+      request = Friendship.find_by(user: current_user, friend: user)
+      if request
+        link_to('Recall friend request', user_friendship_path(id: request.id, user_id: current_user.id, friend_id: user.id), method: :delete, class: 'btn-alert')
+      else
+        link_to('Send friend request', user_friendships_path(user_id: current_user.id, friend_id: user.id, confirmed: false), method: :post, class: 'btn-primary')
+      end
     end
   end
